변경 사항
AdminController, ProductController
- ProductController의 기능 중 상품 관리와 관련된 관리자 기능을 AdminController로 이전

adminHeader.jsp
- 관리자 전용 adminHeader는 새로 만들 예정이므로 기존 코드 삭제 (협의 후 삭제)

detailAdmin.jsp
- Controller 이전으로 인한 하이퍼링크 주소 변경

editAdmin.jsp
- edit.jsp에서 editAdmin.jsp로 이름 변경

header.jsp
- 관리자인 경우 관리자 페이지 링크가 보이도록 설정

/admin/header.jsp
- 관리자 페이지 홈 생성

insertAdmin.jsp
- insert.jsp에서 insertAdmin.jsp로 이름 변경

listAdmin.jsp
- list.jsp에서 listAdmin.jsp로 이름 변경
