+++ semiproject/src/main/webapp/WEB-INF/views/admin/home.jsp
@@ -5,6 +5,10 @@
 	<jsp:param value="관리자 페이지" name="title"/>
 </jsp:include>
 
-<h1>관리자 페이지</h1>
+<div class = "container-1200" style = "height: 1500px">
+
+	<h1>관리자 페이지</h1>
+
+</div>
 
 <jsp:include page="/WEB-INF/views/template/adminFooter.jsp"></jsp:include>
