+++ semiproject/src/main/webapp/WEB-INF/views/product/insertAdmin.jsp
